refactor data access functions
