+++ libs/data-access/src/lib/import-job.ts
@@ -1,5 +1,3 @@
-import { DataClient } from './types';
-
 /**
  * Storage bucket used for DOCX import uploads.
  */
@@ -49,62 +47,3 @@ export interface ImportJobRecord {
   /** ISO timestamp for the latest job update. */
   updated_at?: string;
 }
-
-export const createSignedUploadUrl = async ({
-  client,
-  bucket,
-  path,
-}: {
-  client: DataClient;
-  bucket: string;
-  path: string;
-}) => {
-  const { data, error } = await client.storage
-    .from(bucket)
-    .createSignedUploadUrl(path);
-
-  if (error) {
-    console.error('Failed to create signed upload URL:', error);
-    return null;
-  }
-
-  return data;
-};
-
-export const downloadFromStorage = async ({
-  client,
-  bucket,
-  path,
-}: {
-  client: DataClient;
-  bucket: string;
-  path: string;
-}) => {
-  const { data, error } = await client.storage.from(bucket).download(path);
-
-  if (error) {
-    console.error('Failed to download from storage:', error);
-    return null;
-  }
-
-  return data;
-};
-
-export const storageBucketExists = async ({
-  client,
-  bucket,
-}: {
-  client: DataClient;
-  bucket: string;
-}) => {
-  const { error } = await client.storage.from(bucket).list('', {
-    limit: 1,
-  });
-
-  if (error) {
-    console.error(`Failed to access storage bucket "${bucket}":`, error);
-    return false;
-  }
-
-  return true;
-};
